refactor(core): abstract cart creation and addition logic (#2126)
