+++ .changeset/slimy-trains-rhyme.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": patch
+---
+
+Refactors the add to cart logic to handle some shared functionality like revalidating the tags and setting the cart state.
