+++ core/lib/cart/error.ts
@@ -1,0 +1,6 @@
+export class MissingCartError extends Error {
+  constructor() {
+    super('Cart was not returned in response');
+    this.name = 'MissingCartError';
+  }
+}
